// libs/core/src/identifier-mapping/infrastructure/persistence/entities/connection.orm-entity.d.ts
export declare class ConnectionOrmEntity {
    id: string;
    platformType: string;
    name: string;
    status: string;
    config: Record<string, unknown>;
    credentialsRef: string;
    adapterKey?: string;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=connection.orm-entity.d.ts.map